requirements.txt
google-cloud-storage==2.7.0
google-cloud-bigquery==3.4.2
google-api-core==2.11.0
